+++ packages/expo-modules-core/ios/ReactDelegates/ExpoReactDelegate.swift
@@ -11,6 +11,7 @@ public class ExpoReactDelegate: NSObject {
     self.handlers = handlers
   }
 
+  #if os(iOS) || os(tvOS)
   @objc
   public func createReactRootView(
     moduleName: String,
@@ -32,6 +33,16 @@ public class ExpoReactDelegate: NSObject {
         )
       }()
   }
+  #elseif os(macOS)
+  @objc
+  public func createReactRootView(
+    moduleName: String,
+    initialProperties: [AnyHashable: Any]?,
+    launchOptions: [AnyHashable: Any]?
+  ) -> UIView {
+    return UIView()
+  }
+  #endif
 
   @objc
   public func bundleURL() -> URL? {